Moved LazyLoadingConfigurationException to Exceptions namespace
