+++ src/Microsoft.EntityFrameworkCore.LazyLoading/Microsoft.EntityFrameworkCore.LazyLoading.csproj
@@ -15,8 +15,4 @@
     <PackageReference Include="Microsoft.EntityFrameworkCore" Version="1.1.1" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="Exception\" />
-  </ItemGroup>
-
 </Project>
